error handling 💀

+++ src/Zolupos.Server/Zolupos.Application/Features/Customers/DeleteCustomer.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Zolupos.Application.Common.Interfaces;
 using Zolupos.Application.Common.Wrapper;
+using Zolupos.Shared.Model;
 
 namespace Zolupos.Application.Features.Customers
 {
@@ -21,7 +22,9 @@ namespace Zolupos.Application.Features.Customers
 
         public async Task<ResultWrapper<int>> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
-            var customerToDelete = await _context.Customers.Where(customer => customer.CustomerId == request.Id).FirstAsync();
+            var customerToDelete = await _context.Customers.Where(customer => customer.CustomerId == request.Id).FirstOrDefaultAsync();
+            if (customerToDelete == null)
+                throw new CustomError(Message: "Customer Already exist", Errors: "", StatusCode: System.Net.HttpStatusCode.Conflict);
             _context.Customers.Remove(customerToDelete);
             await _context.SaveChangesAsync();
             return new ResultWrapper<int> { Receive = customerToDelete.CustomerId, Message = ""};
